Remove Top tab
Refactor tabbed column view to use TabView, much simpler internally
